% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ai_new.R
\name{calc.ai_new}
\alias{calc.ai_new}
\title{Determnine Allelic imbalance}
\usage{
calc.ai_new(seg, chrominfo, min.size = 1e+06, cont = 0,
  ploidyByChromosome = TRUE, shrink = TRUE)
}
\arguments{
\item{seg}{segmentation data}

\item{chrominfo}{reference genome version}

\item{min.size}{min.size}

\item{ploidyByChromosome}{option to determine ploidy per chromosome}

\item{shrink}{shrink}
}
\value{
number of Allelic Imbalances
}
\description{
Determnine Allelic imbalance
}
