% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hrd.R
\name{calc.hrd}
\alias{calc.hrd}
\title{Determnine number of loss-off heterosigosity}
\usage{
calc.hrd(seg, nA = 7)
}
\arguments{
\item{seg}{segmentation data}

\item{nA}{column number of copy number of A allele}
}
\value{
number of LOH
}
\description{
Determnine number of loss-off heterosigosity
}
