% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lst.R
\name{calc.lst}
\alias{calc.lst}
\title{Determnine large-scale transitions}
\usage{
calc.lst(seg, chrominfo = chrominfo, nA = 7, chr.arm = "no")
}
\arguments{
\item{seg}{segmentation data}

\item{chrominfo}{reference genome version}

\item{nA}{column number of copy number of A allele}

\item{chr.arm}{option to use chromosome arms defined during segmentation}
}
\value{
number of LSTs
}
\description{
Determnine large-scale transitions
}
