% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.hrd.R
\name{preprocess.hrd}
\alias{preprocess.hrd}
\title{Preprocessing for further analysis}
\usage{
preprocess.hrd(seg)
}
\arguments{
\item{seg}{segmentation data}
}
\value{
preprocessed data
}
\description{
Preprocessing for further analysis
}
