% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.seqz.R
\name{preprocess.seqz}
\alias{preprocess.seqz}
\title{Preprocessing seqz files}
\usage{
preprocess.seqz(seg, ploidy0 = NULL, outputdir = NULL)
}
\arguments{
\item{seg}{A segmentation file}

\item{ploidy0}{ploidy, optional}
}
\value{
Output is
}
\description{
Preprocesses seqz files
}
