% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerocess.hrd.R
\name{prerocess.hrd}
\alias{prerocess.hrd}
\title{Preprocessing for further analysis}
\usage{
prerocess.hrd(seg)
}
\arguments{
\item{seg}{segmentation data}
}
\value{
preprocessed data
}
\description{
Preprocessing for further analysis
}
