% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerocess.seqz.R
\name{prerocess.seqz}
\alias{prerocess.seqz}
\title{Preprocessing seqz files}
\usage{
prerocess.seqz(seg, ploidy0 = NULL)
}
\arguments{
\item{seg}{A segmentation file}

\item{ploidy0}{ploidy, optional}
}
\value{
Output is
}
\description{
Preprocesses seqz files
}
