% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.seg.ai.R
\name{shrink.seg.ai}
\alias{shrink.seg.ai}
\title{unite neighbouring segments if possible}
\usage{
shrink.seg.ai(chr.seg)
}
\arguments{
\item{chr.seg}{segmentation data}
}
\value{
gc normalized copy-number
}
\description{
unite neighbouring segments if possible
}
