% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.seg.ai.wrapper.R
\name{shrink.seg.ai.wrapper}
\alias{shrink.seg.ai.wrapper}
\title{unite neighbouring segments if possible}
\usage{
shrink.seg.ai.wrapper(seg)
}
\arguments{
\item{seg}{segmentation data}
}
\description{
unite neighbouring segments if possible
}
